combsimp.py new file made
// edited by skirpichev
